First refactoring done

+++ android-sdk/src/main/java/com/everypay/sdk/util/EveryPayException.java
@@ -4,14 +4,14 @@ public class EveryPayException extends RuntimeException {
 
     private static final long serialVersionUID = -3213806648620958790L;
 
-    private final int errorCode;
+    private final long errorCode;
 
-    public EveryPayException(final int errorCode, final String message) {
+    public EveryPayException(final long errorCode, final String message) {
         super(message);
         this.errorCode = errorCode;
     }
 
-    public int getErrorCode() {
+    public long getErrorCode() {
         return errorCode;
     }
